requirements.txt
chinese_calendar
lxml
openai>1.0.0
pandas
pyyaml
requests
schedule
pyhandytools
sparkdesk-api==1.3.0
wcferry==39.5.1.0
websocket
pillow
jupyter_client
zhdate
ipykernel
google-generativeai>=0.3.0
dashscope
google-genai
langgraph>=0.4.0
pyyaml>=6.0